# contracts/workflow-http/v1/openapi.yaml
openapi: "3.0.3"
info:
  title: Workflow Runtime HTTP API
  version: "1.5.0"
  description: |
    Current-state freeze of all runtime workflow HTTP endpoints.
    Workflow Assistance V1 is an additive Visit-scoped extension; the exact
    implementation owner SHA is frozen in manifest.json.
servers:
  - url: http://{host}:{port}
    variables:
      host:
        default: 127.0.0.1
      port:
        default: "8989"
security:
  - bearerAuth: []
paths:
...
  /internal/v1/workflow-instances:
    post:
      operationId: createWorkflowInstance
      summary: Create a workflow instance
      tags: [instances]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowInstanceRequest"
      responses:
        "201":
          description: Instance created
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: URL of the created instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateWorkflowInstanceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "425":
          $ref: "#/components/responses/TooEarly"
  /internal/v1/workflow-instances/{workflowInstanceId}:
    get:
      operationId: getWorkflowInstanceDetail
      summary: Get workflow instance detail
      tags: [instances]
      parameters:
        - $ref: "#/components/parameters/WorkflowInstanceId"
      responses:
        "200":
          description: Instance detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowInstanceDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /internal/v1/workflow-instances/{workflowInstanceId}/transitions:
    post:
      operationId: executeTransition
      summary: Execute a workflow transition
      tags: [transitions]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/WorkflowInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteWorkflowTransitionRequest"
      responses:
        "200":
          description: Transition executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteWorkflowTransitionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "425":
          $ref: "#/components/responses/TooEarly"
  /internal/v1/workflow-instances/{workflowInstanceId}/assistance-cases:
    post:
      operationId: requestWorkflowAssistance
      summary: Request Domain Owner assistance for the current Visit
      tags: [assistance]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/WorkflowInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestAssistanceRequest"
      responses:
        "200":
          description: Idempotent replay of an existing Assistance case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistanceCommandResponse"
        "201":
          description: Assistance case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistanceCommandResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "425": { $ref: "#/components/responses/TooEarly" }
  /internal/v1/assistance-cases/{assistanceCaseId}/escalate-to-human:
    post:
      operationId: escalateWorkflowAssistanceToHuman
      summary: Mark an Owner-owned case as requiring external Human input
      tags: [assistance]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/AssistanceCaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalateAssistanceRequest"
      responses:
        "200":
          description: Case escalated to HUMAN_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistanceCommandResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "425": { $ref: "#/components/responses/TooEarly" }
  /internal/v1/assistance-cases/{assistanceCaseId}/resolve:
    post:
      operationId: resolveWorkflowAssistance
      summary: Resolve an open case as the current Domain Owner
      tags: [assistance]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/AssistanceCaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveAssistanceRequest"
      responses:
        "200":
          description: Case resolved; original assignee may transition with the returned version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistanceCommandResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "425": { $ref: "#/components/responses/TooEarly" }
  /internal/v1/assistance-cases/owner-inbox:
    get:
      operationId: listOwnerAssistanceInbox
      summary: List open cases for Domains currently owned by the caller
      tags: [assistance]
      parameters:
        - $ref: "#/components/parameters/BeforeCreatedAt"
        - $ref: "#/components/parameters/BeforeId"
        - $ref: "#/components/parameters/PageLimit"
      responses:
        "200":
          description: Owner assistance page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssistanceCasePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /internal/v1/assistance-cases/human-required:
    get:
      operationId: listHumanRequiredAssistanceInbox
      summary: Minimal cross-Domain HUMAN_REQUIRED projection for coordinators
      description: Does not expose Workflow Context, Submission history, or write capability.
      tags: [assistance]
      parameters:
        - name: beforeEscalatedAt
          in: query
          schema: { type: string, format: date-time }
        - $ref: "#/components/parameters/BeforeId"
        - $ref: "#/components/parameters/PageLimit"
      responses:
        "200":
          description: Human-required assistance page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HumanRequiredAssistancePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /internal/v1/assistance-cases/requested-by-me:
    get:
      operationId: listAssistanceRequestedByMe
      summary: List cases requested by the authenticated caller
      tags: [assistance]
      parameters:
        - $ref: "#/components/parameters/BeforeCreatedAt"
        - $ref: "#/components/parameters/BeforeId"
        - $ref: "#/components/parameters/PageLimit"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/AssistanceCaseStatus" }
      responses:
        "200":
          description: Requester assistance page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssistanceCasePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /internal/v1/assistance-cases/{assistanceCaseId}:
    get:
      operationId: getAssistanceCase
      summary: Get caller-scoped Assistance Case detail
      description: Requesters and current Domain Owners receive full Case detail with the latest Workflow version/current Visit. A global coordinator may read only the minimal HUMAN_REQUIRED projection.
      tags: [assistance]
      parameters:
        - $ref: "#/components/parameters/AssistanceCaseId"
      responses:
        "200":
          description: Assistance case detail
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AssistanceCaseDetail"
                  - $ref: "#/components/schemas/HumanRequiredAssistanceCase"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /internal/v1/workflow-instances/{workflowInstanceId}/timeline:
    get:
      operationId: getTimeline
      summary: Get workflow instance event timeline
      tags: [timeline]
      parameters:
        - $ref: "#/components/parameters/WorkflowInstanceId"
        - name: after
          in: query
          schema:
            type: integer
            description: Event sequence cursor (numeric)
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /internal/v1/workflow-instances/{workflowInstanceId}/submissions:
    get:
      operationId: listSubmissionHistory
      summary: Get workflow instance submission history
      tags: [submissions]
      description: >
        Thin HTTP adapter over the existing list_submission_history query.
        Visibility (full / actor / historical-participant RETURN feedback) and
        pagination (TimeUuidCursor keyset) are owned entirely by the reused
        application query; this endpoint adds no authorization rules.
      parameters:
        - $ref: "#/components/parameters/WorkflowInstanceId"
        - name: afterCreatedAt
          in: query
          schema:
            type: string
            format: date-time
          description: Forward cursor timestamp (must be paired with afterId).
        - name: afterId
          in: query
          schema:
            type: string
            format: uuid
          description: Forward cursor submission id (must be paired with afterCreatedAt).
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Submission history page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionHistoryPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /internal/v1/workflow-instances/domain:
    get:
      operationId: listDomainInstances
      summary: List all instances in a domain (DOMAIN_OWNER only)
      tags: [domain-list]
      parameters:
        - name: domainId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: beforeCreatedAt
          in: query
          schema:
            type: string
            format: date-time
        - name: beforeId
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: definitionKey
          in: query
          schema:
            type: string
        - name: lifecycle
          in: query
          schema:
            type: string
            enum: [active, terminal, all]
        - name: currentNodeKey
          in: query
          schema:
            type: string
        - name: assigneePrincipalId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Paginated domain instance list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainInstancePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /internal/v1/workflow-instances/global:
    get:
      operationId: listGlobalInstances
      summary: List instance summaries across ALL domains (GLOBAL_WORKFLOW_COORDINATOR only)
      description: >-
        Cross-domain read-only view for the formal GLOBAL_WORKFLOW_COORDINATOR
        role. Server-side authorization: callers without the role receive 403
        global_coordinator_required. Returns the same summary projection as the
        domain list (DomainInstanceSummary) — never instance detail or
        submission payload. No domainId parameter; filters mirror the domain
        list minus domainId.
      tags: [global-list]
      parameters:
        - name: beforeCreatedAt
          in: query
          schema:
            type: string
            format: date-time
        - name: beforeId
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: definitionKey
          in: query
          schema:
            type: string
        - name: lifecycle
          in: query
          schema:
            type: string
            enum: [active, terminal, all]
        - name: currentNodeKey
          in: query
          schema:
            type: string
        - name: assigneePrincipalId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [active, cancelled, archived, all]
      responses:
        "200":
          description: Paginated global instance list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainInstancePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /internal/v1/worklists/assigned-to-me:
    get:
      operationId: listAssignedToMe
      summary: List work items assigned to the authenticated actor
      tags: [worklists]
      parameters:
        - name: beforeCreatedAt
          in: query
          schema:
            type: string
            format: date-time
        - name: beforeId
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated worklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorklistPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /internal/v1/worklists/creator-owned-drafts:
    get:
      operationId: listCreatorOwnedDrafts
      summary: List draft instances created by the authenticated actor
      tags: [worklists]
      parameters:
        - name: beforeCreatedAt
          in: query
          schema:
            type: string
            format: date-time
        - name: beforeId
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated creator draft list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorDraftPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  # ── Agent Self-Projection ─────────────────────────────────────────────
  /internal/v1/principals/me:
    put:
      operationId: selfProject
      summary: Self-project authenticated agent identity
      tags: [principals]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Principal projected or already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelfProjectionResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"

  # ── Caller-scoped domain discovery ────────────────────────────────────
  /internal/v1/principals/me/domains:
    get:
      operationId: listMyDomains
      summary: List the caller's own domain memberships
      tags: [principals]
      description: >
        Returns every domain where the verified caller has an enabled
        DOMAIN_OWNER / DOMAIN_MEMBER binding, joined with the domain's
        basic info.  Caller-scoped: the subject comes exclusively from the
        bearer token; no path/query input identifies it.  Disabled
        bindings and disabled domains are excluded.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The caller's domain memberships
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyDomainsList"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── Domain Member Management ──────────────────────────────────────────
  /internal/v1/domains/{domainId}/members:
    get:
      operationId: listDomainMembers
      summary: List DOMAIN_MEMBER bindings for a domain
      tags: [domains]
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: beforeCreatedAt
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: beforeId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Paginated member list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListPage"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /internal/v1/domains/{domainId}/definitions/{definitionId}:
    get:
      operationId: getDomainDefinition
      summary: Get definition detail with all versions
      tags: [definitions]
      parameters:
        - $ref: "#/components/parameters/DomainId"
        - name: definitionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Definition detail with versions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefinitionDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[\\x21-\\x7e]+$"
...
    BeforeCreatedAt:
      name: beforeCreatedAt
      in: query
      schema: { type: string, format: date-time }
...
  schemas:
...
    CreateWorkflowInstanceRequest:
      type: object
      additionalProperties: false
      required: [domainId, definitionVersionId, metadata, contextPayload]
      properties:
        domainId:
          type: string
          format: uuid
        definitionVersionId:
          type: string
          format: uuid
        externalReference:
          type: string
          maxLength: 512
        externalUrl:
          type: string
          format: uri
        metadata:
          type: object
        contextPayload:
          type: object
...
    CreateWorkflowInstanceResponse:
      type: object
      additionalProperties: false
      properties:
        workflowInstanceId:
          type: string
          format: uuid
        workflowStateVersion:
          type: integer
        currentContextRevisionId:
          type: string
          format: uuid
        currentNodeVisitId:
          type: string
          format: uuid
        eventSequence:
          type: integer
      required:
        - workflowInstanceId
        - workflowStateVersion
        - currentContextRevisionId
        - currentNodeVisitId
        - eventSequence
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...